denstrip-package {denstrip}	R Documentation

Overview of the denstrip package

Description

Graphical methods for compactly illustrating and comparing distributions, particularly distributions arising from parameter estimation or prediction.

Details

denstrip implements the density strip for illustrating a single univariate distribution. The darkness of the density strip at a point is proportional to the density at that point. A shortcut function denstrip.normal draws the strip for the given normal distribution.

densregion implements the density region, which illustrates the uncertainty surrounding a continuously-varying quantity as a two-dimensional shaded region with darkness proportional to the density. There are shortcut functions densregion.normal and densregion.survfit for computing and drawing the region for normally-distributed predictions and survival curves, respectively.

sectioned.density implements the sectioned density plots of Cohen and Cohen (2006). These illustrate distributions using occlusion and varying shading. They were developed for summarising data, but can also be used for illustrating known distributions.

vwstrip can be used to draw varying-width strips to illustrate distributions, in a similar manner to the violin plot for summarising data. The width of the strip is proportional to the density. A shortcut function vwstrip.normal draws the strip for the given normal distribution.

bpstrip adapts the box-percentile plot to illustrate a distribution instead of observed data. This strip has width proportional to the probability of a more extreme point.

cistrip implements the popular point and line figure for illustrating point and interval estimates, for example from multiple regression.

These methods are discussed in more detail by Jackson (2008).

Each function is designed to add a graphic to an existing set of plot axes. The plots can be added to either base graphics or lattice panels.

Author(s)

Christopher Jackson <chris.jackson@mrc-bsu.cam.ac.uk>

References

Jackson, C. H. (2008) Displaying uncertainty with shading. The American Statistician, 62(4):340-347.

[Package denstrip version 1.5.4 Index]